+++ app/src/main/java/me/alfredobejarano/brastlewark/model/Gnome.kt
@@ -5,8 +5,8 @@ data class Gnome(
     val name: String = "",
     val thumbnailUrl: String = "",
     val age: Int = 0,
-    val weight: Double = 0.0,
-    val height: Double = 0.0,
+    val weight: Int = 0,
+    val height: Int = 0,
     val hairColor: String = "",
     val professions: List<String> = emptyList(),
     val friends: List<String> = emptyList()
